#517 unify Trees/Trees2 into TreesPlayground with runtime mode switching
Remove duplicate STAMM_SOURCE/STAMM_LAYER constants from ThreeTreeLayer,
import EINZELBAUMX_SOURCE/EINZELBAUMX_LAYER from LoftTreeLayer instead.
Extract radiusMix from URL parsing into a component-level parameter so
both treesFromFeatures() and loftTreesFromFeatures() receive it externally.
Replace separate Trees.tsx/Trees2.tsx with a single TreesPlayground
component that manages useLoft and radiusMix state, handling layer
teardown/rebuild on mode switch.
